requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==26.1.0
audioop-lts==0.2.2
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.10.5
charset-normalizer==3.4.4
discord.py==2.7.1
dotenv==0.9.9
fastapi==0.131.0
frozenlist==1.8.0
idna==3.11
multidict==6.7.1
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
soupsieve==2.8
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.23.0
